requirements.txt
asincio-3.4.3
aiogram-3.13.1
python-dotenv
pandas
pandas openpyxl